Move the `with` wrapper (and friends) into Spectrum.class
For easier access and import. Also to clearly put it on the public API
and not in `internal`.
